Adds temporary filenames to CWL + 90% WIP for WDL
Use the Filename class, you can manually provide an extension to the class and the program will currently just generate a guid as a default, though this behaviour will likely change later. The only problem is it's defined at a Tool level, where this value should probably be mapped at the workflow level.