Changing behaviour for WDL defaults to better align with CWL
This is a pretty big commit, and it's mainly big because it was a little
messy. Basically, instead of binding the default onto the input:

    `String myValue = "myinput"`

The default is now applied on the commandline binding or workflow map:

    # CommandLine binding: allows for nulls
    `if defined($var) then $var else $default`
OR
    # Workflow map: doesn't allow null
    `select_first([$var, defaultvalue])`

This allows you to pass null, closer mimicking CWL's default cascading,
and allows me to remove the hacky `include_in_inputs_if_none` for WDL.
It also makes CPUSelector and MemorySelector subclasses of InputSelector
(which means you get string formatting for free).